+++ release/i18next.amd.withJQuery-1.5.9.js
@@ -1,18 +1,20 @@
-// i18next, v1.5.9pre
+// i18next, v1.5.9
 // Copyright (c)2012 Jan Mühlemann (jamuhl).
 // Distributed under MIT license
 // http://i18next.com
 (function (root, factory) {
     if (typeof exports === 'object') {
 
-        module.exports = factory();
+      var jquery = require('jquery');
+
+      module.exports = factory(jquery);
 
     } else if (typeof define === 'function' && define.amd) {
 
-        define([], factory);
+      define(['jquery'], factory);
 
     } 
-}(this, function () {
+}(this, function ($) {
 
     // add indexOf to non ECMA-262 standard compliant browsers
     if (!Array.prototype.indexOf) {  
@@ -48,8 +50,7 @@
         }
     } 
 
-    var $ = undefined
-        , i18n = {}
+    var i18n = {}
         , resStore = {}
         , currentLng
         , replacementCounter = 0
@@ -67,6 +68,7 @@
         ns: 'translation',
         nsseparator: ':',
         keyseparator: '.',
+        selectorAttr: 'data-i18n',
         debug: false,
         
         resGetPath: 'locales/__lng__/__ns__.json',
@@ -93,12 +95,15 @@
         contextNotFound: ['context_not_found', Math.random()].join(''),
     
         setJqueryExt: true,
+        defaultValueFromContent: true,
         useDataAttrOptions: false,
         cookieExpirationTime: undefined,
         useCookie: true,
+        cookieName: 'i18next',
     
         postProcess: undefined
     };
+    
     function _extend(target, source) {
         if (!source || typeof source === 'function') {
             return target;
@@ -480,7 +485,7 @@
         },
         toLanguages: function(lng) {
             var languages = [];
-            if (lng.indexOf('-') > -1) {
+            if (typeof lng === 'string' && lng.indexOf('-') > -1) {
                 var parts = lng.split('-');
     
                 lng = o.lowerCaseLng ? 
@@ -578,6 +583,73 @@
         return init(cb);
     }
     
+    function setDefaultNamespace(ns) {
+        o.ns.defaultNs = ns;
+    }
+    
+    function loadNamespace(namespace, cb) {
+        loadNamespaces([namespace], cb);
+    }
+    
+    function loadNamespaces(namespaces, cb) {
+        var opts = {
+            dynamicLoad: o.dynamicLoad,
+            resGetPath: o.resGetPath,
+            getAsync: o.getAsync,
+            ns: { namespaces: namespaces, defaultNs: ''} /* new namespaces to load */
+        };
+    
+        // languages to load
+        var lngsToLoad = f.toLanguages(o.lng);
+        if (typeof o.preload === 'string') o.preload = [o.preload];
+        for (var i = 0, l = o.preload.length; i < l; i++) {
+            var pres = f.toLanguages(o.preload[i]);
+            for (var y = 0, len = pres.length; y < len; y++) {
+                if (lngsToLoad.indexOf(pres[y]) < 0) {
+                    lngsToLoad.push(pres[y]);
+                }
+            }
+        }
+    
+        // check if we have to load
+        var lngNeedLoad = [];
+        for (var a = 0, lenA = lngsToLoad.length; a < lenA; a++) {
+            var needLoad = false;
+            var resSet = resStore[lngsToLoad[a]];
+            if (resSet) {
+                for (var b = 0, lenB = namespaces.length; b < lenB; b++) {
+                    if (!resSet[namespaces[b]]) needLoad = true;
+                }
+            } else {
+                needLoad = true;
+            }
+    
+            if (needLoad) lngNeedLoad.push(lngsToLoad[a]);
+        }
+    
+        if (lngNeedLoad.length) {
+            i18n.sync._fetch(lngNeedLoad, opts, function(err, store) {
+                var todo = namespaces.length * lngNeedLoad.length;
+    
+                // load each file individual
+                f.each(namespaces, function(nsIndex, nsValue) {
+                    f.each(lngNeedLoad, function(lngIndex, lngValue) {
+                        resStore[lngValue] = resStore[lngValue] || {};
+                        resStore[lngValue][nsValue] = store[lngValue][nsValue];
+    
+                        todo--; // wait for all done befor callback
+                        if (todo === 0 && cb) {
+                            if (o.useLocalStorage) i18n.sync._storeLocal(resStore);
+                            cb();
+                        }
+                    });
+                });
+            });
+        } else {
+            if (cb) cb();
+        }
+    }
+    
     function setLng(lng, cb) {
         return init({lng: lng}, cb);
     }
@@ -606,20 +678,20 @@
     
             var optionsToUse;
             if (attr === 'html') {
-                optionsToUse = $.extend({ defaultValue: ele.html() }, options);
+                optionsToUse = o.defaultValueFromContent ? $.extend({ defaultValue: ele.html() }, options) : options;
                 ele.html($.t(key, optionsToUse));
             } 
             else if (attr === 'text') {
-                optionsToUse = $.extend({ defaultValue: ele.text() }, options);
+                optionsToUse = o.defaultValueFromContent ? $.extend({ defaultValue: ele.text() }, options) : options;
                 ele.text($.t(key, optionsToUse));
             } else {
-                optionsToUse = $.extend({ defaultValue: ele.attr(attr) }, options);
+                optionsToUse = o.defaultValueFromContent ? $.extend({ defaultValue: ele.attr(attr) }, options) : options;
                 ele.attr(attr, $.t(key, optionsToUse));
             }
         }
     
         function localize(ele, options) {
-            var key = ele.attr('data-i18n');
+            var key = ele.attr(o.selectorAttr);
             if (!key) return;
     
             if (!options && o.useDataAttrOptions === true) {
@@ -648,18 +720,19 @@
                 localize($(this), options);
     
                 // localize childs
-                var elements =  $(this).find('[data-i18n]');
+                var elements =  $(this).find('[' + o.selectorAttr + ']');
                 elements.each(function() { 
                     localize($(this), options);
                 });
             });
         };
     }
+    
     function applyReplacement(str, replacementHash, nestedKey) {
         if (str.indexOf(o.interpolationPrefix) < 0) return str;
     
         f.each(replacementHash, function(key, value) {
-            if (typeof value === 'object') {
+            if (typeof value === 'object' && value !== null) {
                 str = applyReplacement(str, value, nestedKey ? nestedKey + '.' + key : key);
             } else {
                 str = str.replace(new RegExp([o.interpolationPrefix, nestedKey ? nestedKey + '.' + key : key, o.interpolationSuffix].join(''), 'g'), value);
@@ -701,6 +774,8 @@
     function _translate(key, options){
         options = options || {};
     
+        if (!resStore) { return notfound; } // no resStore to translate from
+    
         var optionWithoutCount, translated
           , notfound = options.defaultValue || key
           , lngs = languages;
@@ -718,8 +793,6 @@
                 });
             }
         }
-    
-        if (!resStore) { return notfound; } // no resStore to translate from
     
         var ns = options.ns || o.ns.defaultNs;
         if (key.indexOf(o.nsseparator) > -1) {
@@ -845,7 +918,7 @@
     
         // get from cookie
         if (!detectedLng && typeof document !== 'undefined' && o.useCookie ) {
-            var c = f.cookie.read('i18next');
+            var c = f.cookie.read(o.cookieName);
             if (c) detectedLng = c;
         }
     
@@ -867,7 +940,7 @@
                     }
     
                     if (missingLngs.length > 0) {
-                        sync._fetch(missingLngs, options, function(err, fetched){
+                        sync._fetch(missingLngs, options, function(err, fetched) {
                             f.extend(store, fetched);
                             sync._storeLocal(fetched);
     
@@ -2269,6 +2342,9 @@
     i18n.init = init;
     i18n.setLng = setLng;
     i18n.preload = preload;
+    i18n.loadNamespace = loadNamespace;
+    i18n.loadNamespaces = loadNamespaces;
+    i18n.setDefaultNamespace = setDefaultNamespace;
     i18n.t = translate;
     i18n.translate = translate;
     i18n.detectLanguage = f.detectLanguage;
@@ -2278,7 +2354,10 @@
     i18n.lng = lng;
     i18n.addPostProcessor = addPostProcessor;
     i18n.options = o;
+
+    $.i18n = i18n;
+    $.t = i18n.t;
         
-    return i18n; 
+    return i18n;
 
 }));